+++ lib/generators/katapult/cucumber_features/cucumber_features_generator.rb
@@ -12,10 +12,20 @@ module Katapult
         template 'feature.feature', "features/#{model.name(:variables)}.feature"
       end
 
+      no_tasks do
+        def belongs_tos
+          app_model.get_belongs_tos_for model.name
+        end
+      end
+
       private
 
       def model
         @element
+      end
+
+      def app_model
+        model.application_model
       end
 
     end